requirements.txt
streamlit>=1.16.0
matplotlib>=3.4.3
seaborn>=0.11.2
pandas>=1.3.3
plotly>=5.3.1
squarify>=0.4.3
streamlit_extra